# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Citycard",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/cities/{cityId}/tickets": {
            "get": {
                "tags": [
                    "Ticket"
                ],
                "summary": "Get all tickets of city by cityId",
                "operationId": "2383e84e0021c8b18a988e0982f6e79d",
                "parameters": [
                    {
                        "name": "cityId",
                        "in": "path",
                        "description": "ID of city",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "city_id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "transport_type": {
                                                        "type": "string",
                                                        "example": "bus"
                                                    },
                                                    "ticket_type": {
                                                        "type": "string",
                                                        "example": "standart"
                                                    },
                                                    "price": {
                                                        "type": "number",
                                                        "example": "14"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Ticket"
                ],
                "summary": "Create a new ticket for city by cityId",
                "operationId": "a4e33cba6062cab44483925ba715450a",
                "parameters": [
                    {
                        "name": "cityId",
                        "in": "path",
                        "description": "ID of city",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "transport_type": {
                                                "type": "string"
                                            },
                                            "ticket_type": {
                                                "type": "string"
                                            },
                                            "price": {
                                                "type": "number"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "transport_type": "Автобус",
                                    "ticket_type": "Стандартний",
                                    "price": "14"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "city_id": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "transport_type": {
                                            "type": "string",
                                            "example": "Автобус"
                                        },
                                        "ticket_type": {
                                            "type": "string",
                                            "example": "Стандартний"
                                        },
                                        "price": {
                                            "type": "number",
                                            "example": "14"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The transport_type field format is invalid."
                                        },
                                        "error": {
                                            "properties": {
                                                "transport_type": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The transport_type field format is invalid."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/cities/{cityId}/tickets/{ticketId}": {
            "get": {
                "tags": [
                    "Ticket"
                ],
                "summary": "Get city's ticket by cityId and ticketId",
                "operationId": "c5a5397cf5d5a8926705a89ffa0b40be",
                "parameters": [
                    {
                        "name": "cityId",
                        "in": "path",
                        "description": "ID of city",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "ticketId",
                        "in": "path",
                        "description": "ID of Ticket",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "city_id": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "transport_type": {
                                            "type": "string",
                                            "example": "Автобус"
                                        },
                                        "ticket_type": {
                                            "type": "string",
                                            "example": "Стандартний"
                                        },
                                        "price": {
                                            "type": "number",
                                            "example": "14"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Ticket"
                ],
                "summary": "Update an existing city's ticket by cityId and ticketId",
                "operationId": "042a2601d81189ba1425717a9581bb92",
                "parameters": [
                    {
                        "name": "cityId",
                        "in": "path",
                        "description": "ID of city",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "ticketId",
                        "in": "path",
                        "description": "ID of ticket",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "transport_type": {
                                                "type": "string"
                                            },
                                            "ticket_type": {
                                                "type": "string"
                                            },
                                            "price": {
                                                "type": "number"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "transport_type": "Автобус",
                                    "ticket_type": "Стандартний",
                                    "price": "14"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "city_id": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "transport_type": {
                                            "type": "string",
                                            "example": "Автобус"
                                        },
                                        "ticket_type": {
                                            "type": "string",
                                            "example": "Стандартний"
                                        },
                                        "price": {
                                            "type": "number",
                                            "example": "14"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource not found"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The transport_type field format is invalid."
                                        },
                                        "error": {
                                            "properties": {
                                                "transport_type": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The transport_type field format is invalid."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Ticket"
                ],
                "summary": "Remove ticket by ID",
                "operationId": "8587ad99c556bdb0f0b2c9a65fce4dba",
                "parameters": [
                    {
                        "name": "cityId",
                        "in": "path",
                        "description": "ID of city",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "ticketId",
                        "in": "path",
                        "description": "ID of ticket",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}